project/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
fsspec==2025.9.0
grpcio==1.75.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
openai==1.107.3
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
protobuf==6.32.1
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
PyYAML==6.0.2
qdrant-client==1.15.1
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
